+++ packages/stream_chat_flutter/lib/src/localization/translations.dart
@@ -299,8 +299,9 @@ abstract class Translations {
   /// The text shown for "You"
   String get youText;
 
-  /// The text shown for "Of"
-  String get ofText;
+  /// Gallery footer pagination text
+  String galleryPaginationText(
+      {required int currentPage, required int totalPages});
 
   /// The text shown for "File"
   String get fileText;
@@ -660,7 +661,9 @@ class DefaultTranslations implements Translations {
   String get youText => 'You';
 
   @override
-  String get ofText => 'of';
+  String galleryPaginationText(
+          {required int currentPage, required int totalPages}) =>
+      '${currentPage + 1} of $totalPages';
 
   @override
   String get fileText => 'File';