+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -350,7 +350,9 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
   String get youText => 'You';
 
   @override
-  String get ofText => 'of';
+  String galleryPaginationText(
+          {required int currentPage, required int totalPages}) =>
+      '${currentPage + 1} of $totalPages';
 
   @override
   String get fileText => 'File';